requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
catboost==1.2.1
scikit-learn==1.2.2
Pillow==9.5.0 